requirements.txt
# --------- pytorch --------- #
torch>=2.0.0
# torchvision>=0.15.0
lightning>=2.0.0
torchmetrics>=0.11.4
torchvision>=0.17.2
# --------- hydra --------- #
hydra-core==1.3.2
hydra-colorlog==1.2.0
hydra-optuna-sweeper==1.2.0

# --------- loggers --------- #
wandb==0.16.6
# neptune-client
# mlflow
# comet-ml
# aim>=3.16.2  # no lower than 3.16.2, see https://github.com/aimhubio/aim/issues/2550

# --------- visualizer --------- #
dash>=2.17.0
dash-core-components>=2.0.0
dash-html-components>=2.0.0
dash-table>=5.0.0
importlib_metadata>=7.1.0
importlib_resources>=6.4.0
opencv-python>=4.10.0.82
pillow>=10.3.0
plotly>=5.22.0
plotly-express>=0.4.1
PyYAML>=6.0.1
omegaconf>=2.3.0
dynaconf>=3.2.5
dash-bootstrap-components>=1.6.0
ipython>=8.13.0
numpy>=1.24.4
pandas>=2.0.3

# --------- others --------- #
rootutils       # standardizing the project root setup
pre-commit      # hooks for applying linters on commit
rich            # beautiful text formatting in terminal
pytest          # tests
# sh            # for running bash commands in some tests (linux/macos only)
gradio
